state: Move ACL schema indexes to match Ent
and use constants for table and index names.
